addresses Alex's review suggestions
