+++ isaaclab_arena/embodiments/embodiment_base.py
@@ -9,7 +9,7 @@ from isaaclab.envs import ManagerBasedRLMimicEnv
 from isaaclab.managers.recorder_manager import RecorderManagerBaseCfg
 
 from isaaclab_arena.assets.asset import Asset
-from isaaclab_arena.embodiments.common.mimic_utils import MimicArmMode
+from isaaclab_arena.embodiments.common.mimic_arm_mode import MimicArmMode
 from isaaclab_arena.environments.isaaclab_arena_manager_based_env import IsaacLabArenaManagerBasedRLEnvCfg
 from isaaclab_arena.utils.cameras import make_camera_observation_cfg
 from isaaclab_arena.utils.configclass import combine_configclass_instances
@@ -20,14 +20,14 @@ class EmbodimentBase(Asset):
 
     name: str | None = None
     tags: list[str] = ["embodiment"]
-    mimic_arm_mode: MimicArmMode | None = None
+    default_mimic_arm_mode: MimicArmMode | None = None
 
     def __init__(
         self, enable_cameras: bool = False, initial_pose: Pose | None = None, mimic_arm_mode: MimicArmMode | None = None
     ):
         self.enable_cameras = enable_cameras
         self.initial_pose = initial_pose
-        self.mimic_arm_mode = mimic_arm_mode or type(self).mimic_arm_mode
+        self.mimic_arm_mode = mimic_arm_mode or self.default_mimic_arm_mode
         # These should be filled by the subclass
         self.scene_config: Any | None = None
         self.camera_config: Any | None = None
